requirements.txt
fastapi
uvicorn[standard]
python-multipart

# CORS support for React frontend
fastapi[all]

# Optional but recommended for future API features
pydantic
pydantic[email]

# Database (you said you're using Postgres now, MySQL later)
sqlalchemy
psycopg2-binary
alembic

# If scraping code uses aiohttp or requests, include these:
requests
beautifulsoup4
lxml
aiohttp

pandas
selenium
selenium-stealth
webdriver-manager
openpyxl
pydantic-settings

# Type stubs for mypy
types-requests
types-beautifulsoup4

# Testing
pytest-asyncio
